Code cleanup, add file item state
